Add files via upload
Randomiser is Random choice picker which picks randomly the choices the User provides. 
A Very Good Project!! Sclaled up my Knowledge 
In this project there is a heavy use of setinterval() and setTimeout()
SetInterval() is timing event which runs a function provided to it every 100/200 ms . It keeps running the function every 100/200 ms until its not stopped using CleaInterval() timing event.
setTimeOut() runs a function provided to it after the duration of milli seconds have passed.